docs: update advice for consumers
Correct out-dated README content, and point consumers to the newer
github project Homepage resources.
Reduce the Homepage env table content to better fit the narrow scheme,
and move the bulk of the Authenticating section to the dedicated
sub-page.
Add the auth token example from the Homepage to the auth sub-page, and
consolidate the advice.
Add better linking between the 2 project docs pages and the main README.