feat: better Mongo DB error handling (#160)
* fix: correctly handle mongo DB check
The `connection` test must be awaited to verify DB connectivity.
Exit immediately if the DB connection fails.

* feat: configure mongo timeout
Set maximum time for mongo operations. Configurable via the
`MONGO_TIMEOUT_MS` env var.

* refactor: rename method to avoid clash
Changing from `getPersonIdByCRN` to `findPersonIdByCRN` to better
reflect the DB operation it contains, and to better distinguish this
against other functions with the same name (e.g. in customer resolver).

* chore: improve mongo DB support
Upgrade compose env to use `v7` of mongo DB, in line with CDP.
Allow the DB to be stopped, so it does not persist indefinitely!

* feat: improve log detail, add gateway type

* chore: update dependencies, fix npm audit

* test: move resolver tests to subfolders

* test: add uncovered common customer resolver, and tests for other remaining uncovered areas changed in this PR.